Moved active tab page rendering outside of tab div to resolve parent div boundary issue
